requirements.txt
# Core Framework
Flask==3.1.1
Werkzeug==3.1.3

# IBM Watson Machine Learning
ibm-watson-machine-learning==1.0.368

# Configuration Management
PyYAML==6.0.2

# Document Processing
pdfplumber==0.11.7
openpyxl==3.1.5
python-docx==1.2.0
PyPDF2==3.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.4

# PDF Generation
pdfkit==1.0.0

# HTTP Requests
requests==2.32.4

# Environment Management
python-dotenv==1.1.1

# Development Tools
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2

# Additional Utilities
tqdm==4.67.1
tabulate==0.9.0
colorama==0.4.6
